Se agrega token creation form para creaciones de tokens de manera individual + atomic design methodology para componentes + ajustes varios
