+++ pages/index.js
@@ -1,10 +1,5 @@
 import Head from 'next/head'
-import { Inter } from 'next/font/google'
-import Navbar from '@components/components/Navbar'
-import HeroSection from '@components/components/HeroSection'
-import Footer from '@components/components/Footer'
-
-const inter = Inter({ subsets: ['latin'] })
+import HomeLayout from '@components/components/Templates/HomeLayout'
 
 export default function Home() {
   return (
@@ -15,9 +10,7 @@ export default function Home() {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <main>
-        <Navbar/>
-        <HeroSection/>
-        <Footer/>
+        <HomeLayout/>
       </main>
     </>
   )
