+++ pages/token-creation-form/index.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import TokenCreationLayout from '@components/components/Templates/TokenCreationLayout'
+
+export default function TokenCreationForm() {
+  return (
+    <>
+        <TokenCreationLayout/>
+    </>
+  )
+}
